Optimize Templates

+++ Modules/OnlineAnime/AniLiberty/Controller.cs
@@ -31,7 +31,7 @@ public class AniLibertyController : BaseOnlineController
     [HttpGet]
     [Staticache]
     [Route("lite/aniliberty")]
-    async public Task<ActionResult> Index(string title, int year, string releases, bool rjson = false, bool similar = false, string source = null, string id = null)
+    async public Task<ActionResult> Index(string title, short year, string releases, bool rjson = false, bool similar = false, string source = null, string id = null)
     {
         if (await IsRequestBlocked(rch: true))
             return badInitMsg;
