+++ Modules/OnlineENG/AutoEmbed/Controller.cs
@@ -16,14 +16,14 @@ public class AutoEmbedController : BaseENGController
 
     [HttpGet]
     [Route("lite/autoembed")]
-    public Task<ActionResult> Index(bool checksearch, long id, long tmdb_id, string imdb_id, string title, string original_title, int serial, int s = -1, bool rjson = false)
+    public Task<ActionResult> Index(bool checksearch, long id, long tmdb_id, string imdb_id, string title, string original_title, byte serial, short s = -1, bool rjson = false)
     {
         return ViewTmdb(checksearch, id, tmdb_id, imdb_id, title, original_title, serial, s, rjson, mp4: true, method: "call");
     }
 
     [HttpGet]
     [Route("lite/autoembed/video")]
-    public async Task<ActionResult> Video(long id, int s = -1, int e = -1, bool play = false)
+    public async Task<ActionResult> Video(long id, short s = -1, short e = -1, bool play = false)
     {
         if (PlaywrightBrowser.Status == PlaywrightStatus.disabled)
             return OnError();
@@ -54,7 +54,7 @@ public class AutoEmbedController : BaseENGController
         ));
     }
 
-
+    
     async Task<(string file, List<HeadersModel> headers)> black_magic(string uri)
     {
         if (string.IsNullOrEmpty(uri))
