+++ Modules/OnlineENG/VidSrc/Controller.cs
@@ -19,7 +19,7 @@ public class VidSrcController : BaseENGController
 
     [HttpGet]
     [Route("lite/vidsrc")]
-    public Task<ActionResult> Index(bool checksearch, long id, long tmdb_id, string imdb_id, string title, string original_title, int serial, int s = -1, bool rjson = false)
+    public Task<ActionResult> Index(bool checksearch, long id, long tmdb_id, string imdb_id, string title, string original_title, byte serial, short s = -1, bool rjson = false)
     {
         return ViewTmdb(checksearch, id, tmdb_id, imdb_id, title, original_title, serial, s, rjson, method: "call");
     }
@@ -27,7 +27,7 @@ public class VidSrcController : BaseENGController
     [HttpGet]
     [Route("lite/vidsrc/video")]
     [Route("lite/vidsrc/video.m3u8")]
-    public async Task<ActionResult> Video(long id, string imdb_id, int s = -1, int e = -1, bool play = false)
+    public async Task<ActionResult> Video(long id, string imdb_id, short s = -1, short e = -1, bool play = false)
     {
         if (id == 0)
             return OnError();
