+++ Modules/OnlineGEO/AsiaGe/Controller.cs
@@ -22,7 +22,7 @@ public class AsiaGeController : BaseOnlineController
     [HttpGet]
     [Staticache]
     [Route("lite/asiage")]
-    async public Task<ActionResult> Index(long id, string title, int year, int serial = 0, string href = null, bool similar = false)
+    async public Task<ActionResult> Index(long id, string title, short year, byte serial = 0, string href = null, bool similar = false)
     {
         if (serial == 0)
             return OnError();
