+++ Modules/OnlineRUS/VideoDB/Controller.cs
@@ -20,7 +20,7 @@ public class VideoDBController : BaseOnlineController
 
     [HttpGet]
     [Route("lite/videodb")]
-    async public Task<ActionResult> Index(string uri, string title, string original_title, string t, int s = -1, int sid = -1, bool rjson = false)
+    async public Task<ActionResult> Index(string uri, string title, string original_title, string t, short s = -1, short sid = -1, bool rjson = false)
     {
         string href = DecryptQuery(uri);
