+++ Modules/OnlineUKR/BamBoo/Controller.cs
@@ -15,7 +15,7 @@ public class BamBooController : BaseOnlineController
     [HttpGet]
     [Staticache]
     [Route("lite/bamboo")]
-    async public Task<ActionResult> Index(string title, string original_title, int clarification, int year, int t = -1, string href = null, bool rjson = false, bool similar = false)
+    async public Task<ActionResult> Index(string title, string original_title, byte clarification, short year, short t = -1, string href = null, bool rjson = false, bool similar = false)
     {
         if (await IsRequestBlocked(rch: true))
             return badInitMsg;
